Move inhereted wavelink player class to own module
